pneumonia_prediction/lib/server/requirements.txt
fastapi
uvicorn
tensorflow
numpy
opencv-python
pickle-mixin
python-dotenv
google-generativeai
python-multipart

